import-mod: Prefer double bracket bash extension usage
This is to maintain the sequence of its usage and not to mix the use
of POSIX and bash extenstion for the brackets. Since this is a bash
script, we will never run into compatibility issues. Moreover, the
bash extension is newer and has more advantages comparatively.

Check out: https://unix.stackexchange.com/questions/49007/when-should-i-use-vs-in-bash-single-vs-double-brackets

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>